chore: apply suggestions
